fix: tabs back navigation
