--- src/app/(settings)/_layout.tsx
@@ -1,5 +1,0 @@
-import { Stack } from "expo-router";
-
-export default function SettingsLayout() {
-  return <Stack screenOptions={{ headerShown: false }} />;
-}
