+++ src/app/(tabs)/(password-manager)/_layout.tsx
