+++ src/app/(tabs)/(password-manager)/index.tsx
