+++ src/app/(tabs)/(password-manager)/password-details.tsx
