+++ src/app/(tabs)/(settings)/_layout.tsx
@@ -1,0 +1,9 @@
+import { Stack } from "expo-router";
+
+export default function SettingsLayout() {
+  return (
+    <Stack screenOptions={{ headerShown: false }}>
+      <Stack.Screen name="settings"></Stack.Screen>
+    </Stack>
+  );
+}
